# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Użytkownik zarejestrowany pomyślnie"
          },
          "400": {
            "description": "Nieprawidłowe dane wejściowe"
          },
          "409": {
            "description": "Email już jest w użyciu"
          },
          "429": {
            "description": "Zbyt wiele prób rejestracji. Spróbuj później."
          }
        },
        "summary": "Rejestracja nowego użytkownika (sesja w cookie httpOnly + token w odpowiedzi)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Zalogowano pomyślnie"
          },
          "401": {
            "description": "Nieprawidłowy email lub hasło"
          },
          "429": {
            "description": "Zbyt wiele prób logowania. Spróbuj później."
          }
        },
        "summary": "Logowanie za pomocą email i hasła",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/google": {
      "post": {
        "operationId": "AuthController_googleLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Zalogowano przez Google pomyślnie"
          },
          "401": {
            "description": "Nieprawidłowy token Google"
          }
        },
        "summary": "Logowanie za pomocą Google",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/session": {
      "post": {
        "operationId": "AuthController_session",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ping sesji – zwraca id zalogowanego użytkownika (do bootstrapu SPA)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change own password (needs the current one); other sessions are logged out, this one gets a fresh cookie",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Jeśli email istnieje, wysłano link resetujący"
          },
          "429": {
            "description": "Zbyt wiele prób. Spróbuj później."
          }
        },
        "summary": "Żądanie resetu hasła (wysyła e-mail z linkiem)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Hasło zmienione"
          },
          "400": {
            "description": "Token nieprawidłowy lub wygasł"
          }
        },
        "summary": "Ustawienie nowego hasła na podstawie tokenu z e-maila",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/allegro/status": {
      "get": {
        "operationId": "AllegroController_status",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Is the integration configured / is this user connected",
        "tags": [
          "allegro"
        ]
      }
    },
    "/api/allegro/auth-url": {
      "get": {
        "operationId": "AllegroController_authUrl",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Allegro OAuth authorization URL (redirect the browser there)",
        "tags": [
          "allegro"
        ]
      }
    },
    "/api/allegro/callback": {
      "post": {
        "operationId": "AllegroController_callback",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CallbackDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Exchange the OAuth code for tokens",
        "tags": [
          "allegro"
        ]
      }
    },
    "/api/allegro/connection": {
      "delete": {
        "operationId": "AllegroController_disconnect",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Disconnect the Allegro account (deletes stored tokens)",
        "tags": [
          "allegro"
        ]
      }
    },
    "/api/allegro/offers": {
      "get": {
        "operationId": "AllegroController_offers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the seller's active offers",
        "tags": [
          "allegro"
        ]
      }
    },
    "/api/allegro/offers/{offerId}/import": {
      "post": {
        "operationId": "AllegroController_importOffer",
        "parameters": [
          {
            "name": "offerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Import the offer's main photo as a new product image",
        "tags": [
          "allegro"
        ]
      }
    },
    "/api/allegro/offers/{offerId}/publish": {
      "post": {
        "operationId": "AllegroController_publish",
        "parameters": [
          {
            "name": "offerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Publish a generated graphic to the offer (as main photo or appended to the gallery)",
        "tags": [
          "allegro"
        ]
      }
    },
...
    "/api/descriptions/{imageId}": {
      "get": {
        "operationId": "DescriptionsController_get",
        "parameters": [
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Offer copy (title, HTML description, keywords) attached to an uploaded photo",
        "tags": [
          "descriptions"
        ]
      },
      "post": {
        "operationId": "DescriptionsController_create",
        "parameters": [
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDescriptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Write the offer copy from the seller notes – free bonus of a finished graphic; writing again spends a prompt edit",
        "tags": [
          "descriptions"
        ]
      },
      "patch": {
        "operationId": "DescriptionsController_update",
        "parameters": [
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDescriptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Manual edit of the offer copy (unlimited, sanitized to Allegro HTML)",
        "tags": [
          "descriptions"
        ]
      }
    },
    "/api/descriptions/{imageId}/refine": {
      "post": {
        "operationId": "DescriptionsController_refine",
        "parameters": [
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefineDescriptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "AI rewrite following an instruction – spends one prompt edit (5 free per photo, then packs)",
        "tags": [
          "descriptions"
        ]
      }
    },
    "/api/descriptions/{imageId}/edit-packs": {
      "post": {
        "operationId": "DescriptionsController_buyEditPack",
        "parameters": [
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buy a pack of 15 prompt edits for this photo (1 credit)",
        "tags": [
          "descriptions"
        ]
      }
    }
  },
  "info": {
    "title": "AllGrafika API",
    "description": "API for generating product graphics",
    "version": "1.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "StrongPassword123!"
          },
          "name": {
            "type": "string",
            "example": "Jan Kowalski"
          },
          "acceptedTerms": {
            "type": "boolean",
            "example": true,
            "description": "Akceptacja regulaminu i polityki prywatności"
          }
        },
        "required": [
          "email",
          "password",
          "acceptedTerms"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "StrongPassword123!"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "GoogleLoginDto": {
        "type": "object",
        "properties": {
          "googleToken": {
            "type": "string",
            "description": "Google ID token from the frontend",
            "example": "eyJhbGciOiJSUzI1NiIsImtpZCI6IjE4MmU0M..."
          },
          "acceptedTerms": {
            "type": "boolean",
            "description": "Akceptacja regulaminu (wymagana tylko dla nowych kont)"
          }
        },
        "required": [
          "googleToken"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string",
            "example": "NewStrongPassword123!"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "token",
          "password"
        ]
      },
...
      "CallbackDto": {
        "type": "object",
        "properties": {}
      },
      "PublishDto": {
        "type": "object",
        "properties": {}
      },
      "CreateDescriptionDto": {
        "type": "object",
        "properties": {
          "notes": {
            "type": "string",
            "description": "Informacje od sprzedawcy: cechy, wymiary, materiał, zawartość zestawu, przewagi"
          }
        }
      },
      "UpdateDescriptionDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 75,
            "description": "Tytuł oferty (limit Allegro: 75 znaków)"
          },
          "body": {
            "type": "string",
            "maxLength": 20000,
            "description": "Opis w HTML (dozwolone: h2, p, ul, ol, li, b)"
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "body"
        ]
      },
      "RefineDescriptionDto": {
        "type": "object",
        "properties": {
          "instruction": {
            "type": "string",
            "description": "Co zmienić, np. \"skróć o połowę i dodaj sekcję o gwarancji\""
          }
        },
        "required": [
          "instruction"
        ]
      }
    }
  }
}